+++ datadog_checks_base/datadog_checks/base/checks/windows/perf_counters/connection.py
@@ -22,7 +22,7 @@ class NetworkResources:
             name = resource.lpRemoteName
             if name not in self.__resources:
                 # https://docs.microsoft.com/en-us/windows/win32/api/winnetwk/nf-winnetwk-wnetaddconnection2a
-                # http://timgolden.me.uk/pywin32-docs/win32wnet__WNetAddConnection2_meth.html
+                # https://mhammond.github.io/pywin32/win32wnet__WNetAddConnection2_meth.html
                 win32wnet.WNetAddConnection2(resource, password, username, 0)
 
             self.__resources[name] += 1
@@ -38,7 +38,7 @@ class NetworkResources:
                 del self.__resources[name]
 
                 # https://docs.microsoft.com/en-us/windows/win32/api/winnetwk/nf-winnetwk-wnetcancelconnection2a
-                # http://timgolden.me.uk/pywin32-docs/win32wnet__WNetCancelConnection2_meth.html
+                # https://mhammond.github.io/pywin32/win32wnet__WNetCancelConnection2_meth.html
                 win32wnet.WNetCancelConnection2(name, 0, 1)
 
 
@@ -75,7 +75,7 @@ class Connection:
                 server = f'{server}.ipv6-literal.net'
 
             # https://docs.microsoft.com/en-us/windows/win32/api/winnetwk/ns-winnetwk-netresourcea
-            # http://timgolden.me.uk/pywin32-docs/PyNETRESOURCE.html
+            # https://mhammond.github.io/pywin32/PyNETRESOURCE.html
             self.network_resource = win32wnet.NETRESOURCE()
             self.network_resource.lpRemoteName = fr'\\{server}'
 
@@ -90,12 +90,12 @@ class Connection:
             self.network_resources.add(self.network_resource, self.username, self.password)
 
         # https://docs.microsoft.com/en-us/windows/win32/api/pdh/nf-pdh-pdhopenquerya
-        # http://timgolden.me.uk/pywin32-docs/win32pdh__OpenQuery_meth.html
+        # https://mhammond.github.io/pywin32/win32pdh__OpenQuery_meth.html
         self.__query_handle = win32pdh.OpenQuery()
 
     def disconnect(self):
         # https://docs.microsoft.com/en-us/windows/win32/api/pdh/nf-pdh-pdhclosequery
-        # http://timgolden.me.uk/pywin32-docs/win32pdh__CloseQuery_meth.html
+        # https://mhammond.github.io/pywin32/win32pdh__CloseQuery_meth.html
         win32pdh.CloseQuery(self.__query_handle)
 
         if self.network_resource is not None: