+++ datadog_checks_base/datadog_checks/base/utils/tracing.py
@@ -2,8 +2,7 @@
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 import functools
-
-from ddtrace import tracer
+import os
 
 from ..config import is_affirmative
 
@@ -12,6 +11,9 @@ try:
 except ImportError:
     # Integration Tracing is only available with Agent 6
     datadog_agent = None
+
+
+EXCLUDED_MODULES = ['threading']
 
 
 def traced(fn):
@@ -40,8 +42,50 @@ def traced(fn):
         integration_tracing = is_affirmative(datadog_agent.get_config('integration_tracing'))
 
         if integration_tracing and trace_check:
-            with tracer.trace(self.name, service='integrations-tracing', resource=fn.__name__):
-                return fn(self, *args, **kwargs)
+            try:
+                from ddtrace import patch_all, tracer
+
+                patch_all()
+                with tracer.trace(self.name, service='integrations-tracing', resource=fn.__name__):
+                    return fn(self, *args, **kwargs)
+            except Exception:
+                pass
         return fn(self, *args, **kwargs)
 
     return traced_wrapper
+
+
+def tracing_method(f, tracer):
+    @functools.wraps(f)
+    def wrapper(*args, **kwargs):
+        with tracer.trace(f.__name__, resource=f.__name__):
+            return f(*args, **kwargs)
+
+    return wrapper
+
+
+def traced_class(cls):
+    if os.getenv('DDEV_TRACE_ENABLED', 'false') == 'true':
+        try:
+            from ddtrace import patch_all, tracer
+
+            patch_all()
+
+            def decorate(cls):
+                for attr in cls.__dict__:
+                    # Ignoring staticmethod and classmethod because they don't need cls in args
+                    if (
+                        callable(getattr(cls, attr))
+                        and not isinstance(cls.__dict__[attr], staticmethod)
+                        and not isinstance(cls.__dict__[attr], classmethod)
+                        # Get rid of SnmpCheck._thread_factory and related
+                        and getattr(getattr(cls, attr), '__module__', 'threading') not in EXCLUDED_MODULES
+                    ):
+                        setattr(cls, attr, tracing_method(getattr(cls, attr), tracer))
+                return cls
+
+            return decorate(cls)
+        except Exception:
+            pass
+
+    return cls