+++ elastic/datadog_checks/elastic/config_models/instance.py
@@ -44,6 +44,7 @@ class InstanceConfig(BaseModel):
     cat_allocation_stats: Optional[bool]
     cluster_stats: Optional[bool]
     connect_timeout: Optional[float]
+    detailed_index_stats: Optional[bool]
     disable_generic_tags: Optional[bool]
     disable_legacy_cluster_tag: Optional[bool]
     empty_default_hostname: Optional[bool]