+++ nginx/datadog_checks/nginx/nginx.py
@@ -12,7 +12,8 @@ from six.moves.urllib.parse import urlparse
 from datadog_checks.base import AgentCheck, ConfigurationError, to_native_string
 from datadog_checks.base.utils.time import get_timestamp
 
-from .metrics import METRICS_SEND_AS_COUNT, VTS_METRIC_MAP
+from .const import PLUS_API_ENDPOINTS, PLUS_API_STREAM_ENDPOINTS, TAGGED_KEYS
+from .metrics import COUNT_METRICS, METRICS_SEND_AS_COUNT, VTS_METRIC_MAP
 
 if PY3:
     long = int
@@ -25,39 +26,6 @@ else:
 
     def fromisoformat(ts):
         return datetime.strptime(ts, "%Y-%m-%dT%H:%M:%S")
-
-
-PLUS_API_ENDPOINTS = {
-    "nginx": [],
-    "http/requests": ["requests"],
-    "http/server_zones": ["server_zones"],
-    "http/upstreams": ["upstreams"],
-    "http/caches": ["caches"],
-    "processes": ["processes"],
-    "connections": ["connections"],
-    "ssl": ["ssl"],
-    "slabs": ["slabs"],
-}
-
-PLUS_API_STREAM_ENDPOINTS = {
-    "stream/server_zones": ["stream", "server_zones"],
-    "stream/upstreams": ["stream", "upstreams"],
-}
-
-PLUS_API_V3_STREAM_ENDPOINTS = {
-    "stream/zone_sync": ["stream", "zone_sync"],
-}
-
-TAGGED_KEYS = {
-    'caches': 'cache',
-    'server_zones': 'server_zone',
-    'serverZones': 'server_zone',  # VTS
-    'upstreams': 'upstream',
-    'upstreamZones': 'upstream',  # VTS
-    'slabs': 'slab',
-    'slots': 'slot',
-    'zones': 'zone',
-}
 
 
 class Nginx(AgentCheck):
@@ -92,80 +60,92 @@ class Nginx(AgentCheck):
 
     def check(self, _):
         if not self.use_plus_api:
-            response, content_type, version = self._get_data()
-            # for unpaid versions
-            self._set_version_metadata(version)
-
-            self.log.debug("Nginx status `response`: %s", response)
-            self.log.debug("Nginx status `content_type`: %s", content_type)
-
-            if content_type.startswith('application/json'):
-                metrics = self.parse_json(response, self.custom_tags)
-            else:
-                metrics = self.parse_text(response, self.custom_tags)
+            metrics = self.collect_unit_metrics()
         else:
-            metrics = []
-            self._perform_service_check('{}/{}'.format(self.url, self.plus_api_version))
+            metrics = self.collect_plus_metrics()
 
-            # These are all the endpoints we have to call to get the same data as we did with the old API
-            # since we can't get everything in one place anymore.
-
-            if self.use_plus_api_stream:
-                plus_api_chain_list = chain(iteritems(PLUS_API_ENDPOINTS), self._get_plus_api_stream_endpoints())
-            else:
-                plus_api_chain_list = chain(iteritems(PLUS_API_ENDPOINTS))
-
-            for endpoint, nest in plus_api_chain_list:
-                response = self._get_plus_api_data(endpoint, nest)
-
-                if endpoint == 'nginx':
-                    try:
-                        if isinstance(response, dict):
-                            version_plus = response.get('version')
-                        else:
-                            version_plus = json.loads(response).get('version')
-                        self._set_version_metadata(version_plus)
-                    except Exception as e:
-                        self.log.debug("Couldn't submit nginx version: %s", e)
-
-                self.log.debug("Nginx Plus API version %s `response`: %s", self.plus_api_version, response)
-                metrics.extend(self.parse_json(response, self.custom_tags))
-
-        funcs = {'gauge': self.gauge, 'rate': self.rate, 'count': self.monotonic_count}
+        metric_submission_funcs = {'gauge': self.gauge, 'rate': self.rate, 'count': self.monotonic_count}
         conn = None
         handled = None
 
         for row in metrics:
             try:
                 name, value, tags, metric_type = row
-
-                # Translate metrics received from VTS
                 if self.use_vts:
-                    # Requests per second
-                    if name == 'nginx.connections.handled':
-                        handled = value
-                    if name == 'nginx.connections.accepted':
-                        conn = value
-                        self.rate('nginx.net.conn_opened_per_s', conn, tags)
-                    if handled is not None and conn is not None:
-                        self.rate('nginx.net.conn_dropped_per_s', conn - handled, tags)
-                        handled = None
-                        conn = None
-                    if name == 'nginx.connections.requests':
-                        self.rate('nginx.net.request_per_s', value, tags)
-
-                    name = VTS_METRIC_MAP.get(name)
+                    name, handled, conn = self._translate_from_vts(name, value, tags, handled, conn)
                     if name is None:
                         continue
 
-                if name in METRICS_SEND_AS_COUNT:
-                    func_count = funcs['count']
-                    func_count(name + "_count", value, tags)
-                func = funcs[metric_type]
-                func(name, value, tags)
+                if name in COUNT_METRICS:
+                    func_count = metric_submission_funcs['count']
+                    func_count(name, value, tags)
+                else:
+                    if name in METRICS_SEND_AS_COUNT:
+                        func_count = metric_submission_funcs['count']
+                        func_count(name + "_count", value, tags)
+
+                    func = metric_submission_funcs[metric_type]
+                    func(name, value, tags)
 
             except Exception as e:
                 self.log.error('Could not submit metric: %s: %s', repr(row), e)
+
+    def collect_plus_metrics(self):
+        metrics = []
+        self._perform_service_check('{}/{}'.format(self.url, self.plus_api_version))
+
+        # These are all the endpoints we have to call to get the same data as we did with the old API
+        # since we can't get everything in one place anymore.
+
+        plus_api_chain_list = self._get_all_plus_api_endpoints()
+
+        for endpoint, nest in plus_api_chain_list:
+            response = self._get_plus_api_data(endpoint, nest)
+
+            if endpoint == 'nginx':
+                try:
+                    if isinstance(response, dict):
+                        version_plus = response.get('version')
+                    else:
+                        version_plus = json.loads(response).get('version')
+                    self._set_version_metadata(version_plus)
+                except Exception as e:
+                    self.log.debug("Couldn't submit nginx version: %s", e)
+
+            self.log.debug("Nginx Plus API version %s `response`: %s", self.plus_api_version, response)
+            metrics.extend(self.parse_json(response, self.custom_tags))
+        return metrics
+
+    def collect_unit_metrics(self):
+        response, content_type, version = self._get_data()
+        # for unpaid versions
+        self._set_version_metadata(version)
+
+        self.log.debug("Nginx status `response`: %s", response)
+        self.log.debug("Nginx status `content_type`: %s", content_type)
+
+        if content_type.startswith('application/json'):
+            metrics = self.parse_json(response, self.custom_tags)
+        else:
+            metrics = self.parse_text(response, self.custom_tags)
+        return metrics
+
+    def _translate_from_vts(self, name, value, tags, handled, conn):
+        # Requests per second
+        if name == 'nginx.connections.handled':
+            handled = value
+        if name == 'nginx.connections.accepted':
+            conn = value
+            self.rate('nginx.net.conn_opened_per_s', conn, tags)
+        if handled is not None and conn is not None:
+            self.rate('nginx.net.conn_dropped_per_s', conn - handled, tags)
+            handled = None
+            conn = None
+        if name == 'nginx.connections.requests':
+            self.rate('nginx.net.request_per_s', value, tags)
+
+        name = VTS_METRIC_MAP.get(name)
+        return name, handled, conn
 
     def _get_data(self):
         r = self._perform_service_check(self.url)
@@ -205,10 +185,22 @@ class Nginx(AgentCheck):
 
         return {keys[0]: self._nest_payload(keys[1:], payload)}
 
-    def _get_plus_api_stream_endpoints(self):
-        endpoints = iteritems(PLUS_API_STREAM_ENDPOINTS)
-        if int(self.plus_api_version) >= 3:
-            endpoints = chain(endpoints, iteritems(PLUS_API_V3_STREAM_ENDPOINTS))
+    def _get_plus_api_endpoints(self, use_stream=False):
+        endpoints = iteritems({})
+
+        available_plus_endpoints = PLUS_API_STREAM_ENDPOINTS if use_stream else PLUS_API_ENDPOINTS
+
+        for earliest_version, new_endpoints in available_plus_endpoints.items():
+            if int(self.plus_api_version) >= int(earliest_version):
+                endpoints = chain(endpoints, iteritems(new_endpoints))
+        return endpoints
+
+    def _get_all_plus_api_endpoints(self):
+        endpoints = self._get_plus_api_endpoints()
+
+        if self.use_plus_api_stream:
+            endpoints = chain(endpoints, self._get_plus_api_endpoints(use_stream=True))
+
         return endpoints
 
     def _get_plus_api_data(self, endpoint, nest):
@@ -222,7 +214,7 @@ class Nginx(AgentCheck):
             r = self._perform_request(url)
             payload = self._nest_payload(nest, r.json())
         except Exception as e:
-            if endpoint in PLUS_API_STREAM_ENDPOINTS or endpoint in PLUS_API_V3_STREAM_ENDPOINTS:
+            if endpoint in PLUS_API_STREAM_ENDPOINTS.values():
                 self.log.warning("Stream may not be initialized. Error querying %s metrics at %s: %s", endpoint, url, e)
             else:
                 self.log.exception("Error querying %s metrics at %s: %s", endpoint, url, e)
@@ -297,7 +289,6 @@ class Nginx(AgentCheck):
         Recursively flattens the nginx json object. Returns the following: [(metric_name, value, tags)]
         """
         output = []
-
         if isinstance(val, dict):
             # Pull out the server as a tag instead of trying to read as a metric
             if 'server' in val and val['server']: