+++ snmp/tests/conftest.py
@@ -4,6 +4,7 @@
 
 import os
 import shutil
+import socket
 from copy import deepcopy
 
 import pytest
@@ -16,11 +17,8 @@ from datadog_checks.dev.docker import get_container_ip
 from .common import (
     COMPOSE_DIR,
     PORT,
-    SCALAR_OBJECTS,
-    SCALAR_OBJECTS_WITH_TAGS,
     SNMP_CONTAINER_NAME,
     SNMP_LISTENER_ENV,
-    TABULAR_OBJECTS,
     TOX_ENV_NAME,
     generate_container_instance_config,
 )
@@ -57,8 +55,14 @@ def dd_environment():
                     '{}:/etc/datadog-agent/datadog.yaml'.format(create_datadog_conf_file(tmp_dir))
                 ]
             else:
-                instance_config = generate_container_instance_config(
-                    SCALAR_OBJECTS + SCALAR_OBJECTS_WITH_TAGS + TABULAR_OBJECTS
+                instance_config = generate_container_instance_config([])
+                instance_config['init_config'].update(
+                    {
+                        'loader': 'core',
+                        'use_device_id_as_hostname': True,
+                        # use hostname as namespace to create different device for each user
+                        'namespace': socket.gethostname(),
+                    }
                 )
             yield instance_config, new_e2e_metadata