+++ sqlserver/tests/test_activity.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
 import json
 import os
 from concurrent.futures.thread import ThreadPoolExecutor
@@ -5,6 +8,7 @@ from copy import copy
 
 import mock
 import pytest
+from dateutil import parser
 
 from datadog_checks.base.utils.db.utils import DBMAsyncJob, default_json_event_encoding
 from datadog_checks.sqlserver import SQLServer
@@ -63,7 +67,7 @@ def test_collect_activity_windows(aggregator, instance_docker, dd_run_check, ins
 
 def _run_test_collect_activity(aggregator, instance_docker, dd_run_check, dbm_instance):
     check = SQLServer(CHECK_NAME, {}, [dbm_instance])
-    query = "select * from things"
+    query = "SELECT * FROM ϑings"
     bob_conn = _get_conn_for_user(instance_docker, "bob")
     with bob_conn.cursor() as cursor:
         cursor.execute("USE {}".format("datadog_test"))
@@ -106,7 +110,13 @@ def _run_test_collect_activity(aggregator, instance_docker, dd_run_check, dbm_in
     assert bobs_row['database_name'] == "datadog_test", "incorrect database_name"
     assert bobs_row['session_status'] == "sleeping", "incorrect session_status"
     assert bobs_row['id'], "missing session id"
+    assert bobs_row['now'], "missing current timestamp"
     assert bobs_row['transaction_begin_time'], "missing tx begin time"
+
+    # assert that the tx begin time is being collected as an ISO timestamp with TZ info
+    assert parser.isoparse(bobs_row['transaction_begin_time']).tzinfo, "tx begin timestamp not formatted correctly"
+    # assert that the current timestamp is being collected as an ISO timestamp with TZ info
+    assert parser.isoparse(bobs_row['now']).tzinfo, "current timestamp not formatted correctly"
 
     assert len(first['sqlserver_connections']) > 0
     b_conn = None
@@ -119,8 +129,9 @@ def _run_test_collect_activity(aggregator, instance_docker, dd_run_check, dbm_in
 
     # internal debug metrics
     aggregator.assert_metric(
-        "dd.sqlserver.activity.collect_activity.time",
-        tags=['agent_hostname:stubbed.hostname'] + _expected_dbm_instance_tags(dbm_instance),
+        "dd.sqlserver.operation.time",
+        tags=['agent_hostname:stubbed.hostname', 'operation:collect_activity']
+        + _expected_dbm_instance_tags(dbm_instance),
     )
 
     # finally, on the second iteration, only bob's transaction is still open