+++ datadog_checks_base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG - datadog_checks_base
 
+## 23.5.0 / 2021-12-08
+
+* [Added] Add decorator for tracking execution statistics of check methods. See [#10809](https://github.com/DataDog/integrations-core/pull/10809).
+* [Added] Add detailed trace to all integrations. See [#10679](https://github.com/DataDog/integrations-core/pull/10679).
+* [Fixed] Import ddtrace only when needed. See [#10800](https://github.com/DataDog/integrations-core/pull/10800).
+
+## 23.4.0 / 2021-11-30
+
+* [Added] [OpenMetricsV2] Support custom transformers by regex matching metric names. See [#10753](https://github.com/DataDog/integrations-core/pull/10753).
+* [Fixed] Bump cachetools. See [#10742](https://github.com/DataDog/integrations-core/pull/10742).
+* [Fixed] Bump redis dependency. See [#9383](https://github.com/DataDog/integrations-core/pull/9383).
+
 ## 23.3.2 / 2021-11-23
 
 * [Fixed] [PerfCountersBaseCheck] Improve logging when expected counters are not found. See [#10701](https://github.com/DataDog/integrations-core/pull/10701).